+++ src/actions/include-main.ts
@@ -1,0 +1,9 @@
+import './base';
+import './operator';
+import './motion';
+import './textobject';
+
+// commands
+import './commands/insert';
+import './commands/actions';
+import './commands/commandLine';
