+++ src/actions/include-plugins.ts
@@ -1,0 +1,7 @@
+// plugin
+import './plugins/camelCaseMotion';
+import './plugins/easymotion/easymotion.cmd';
+import './plugins/easymotion/registerMoveActions';
+import './plugins/sneak';
+import './plugins/replaceWithRegister';
+import './plugins/surround';
