+++ src/configuration/validators/inputMethodSwitcherValidator.ts
@@ -1,8 +1,8 @@
 import { IConfigurationValidator, ValidatorResults } from '../iconfigurationValidator';
 import { IConfiguration } from '../iconfiguration';
-import { promisify } from 'util';
-import { exists } from 'fs';
+import { existsAsync } from '../../util/fs';
 import { Globals } from '../../globals';
+import { configurationValidator } from '../configurationValidator';
 
 export class InputMethodSwitcherConfigurationValidator implements IConfigurationValidator {
   async validate(config: IConfiguration): Promise<ValidatorResults> {
@@ -27,7 +27,7 @@ export class InputMethodSwitcherConfigurationValidator implements IConfiguration
         level: 'error',
         message: 'vim.autoSwitchInputMethod.obtainIMCmd is empty.',
       });
-    } else if (!(await promisify(exists)(this.getRawCmd(inputMethodConfig.obtainIMCmd)))) {
+    } else if (!(await existsAsync(this.getRawCmd(inputMethodConfig.obtainIMCmd)))) {
       result.append({
         level: 'error',
         message: `Unable to find ${inputMethodConfig.obtainIMCmd}. Check your 'vim.autoSwitchInputMethod.obtainIMCmd' in VSCode setting.`,
@@ -39,7 +39,7 @@ export class InputMethodSwitcherConfigurationValidator implements IConfiguration
         level: 'error',
         message: 'vim.autoSwitchInputMethod.defaultIM is empty.',
       });
-    } else if (!(await promisify(exists)(this.getRawCmd(inputMethodConfig.switchIMCmd)))) {
+    } else if (!(await existsAsync(this.getRawCmd(inputMethodConfig.switchIMCmd)))) {
       result.append({
         level: 'error',
         message: `Unable to find ${inputMethodConfig.switchIMCmd}. Check your 'vim.autoSwitchInputMethod.switchIMCmd' in VSCode setting.`,
@@ -57,3 +57,5 @@ export class InputMethodSwitcherConfigurationValidator implements IConfiguration
     return cmd.split(' ')[0];
   }
 }
+
+configurationValidator.registerValidator(new InputMethodSwitcherConfigurationValidator());
