+++ src/configuration/validators/neovimValidator.ts
@@ -3,6 +3,7 @@ import { IConfigurationValidator, ValidatorResults } from '../iconfigurationVali
 import { execFileSync } from 'child_process';
 import * as path from 'path';
 import { existsSync } from 'fs';
+import { configurationValidator } from '../configurationValidator';
 
 export class NeovimValidator implements IConfigurationValidator {
   validate(config: IConfiguration): Promise<ValidatorResults> {
@@ -49,3 +50,5 @@ export class NeovimValidator implements IConfigurationValidator {
     config.enableNeovim = false;
   }
 }
+
+configurationValidator.registerValidator(new NeovimValidator());
