+++ src/configuration/validators/vimrcValidator.ts
@@ -1,5 +1,7 @@
 import { IConfiguration } from '../iconfiguration';
 import { IConfigurationValidator, ValidatorResults } from '../iconfigurationValidator';
+import { vimrc } from '../vimrc';
+import { configurationValidator } from '../configurationValidator';
 
 export class VimrcValidator implements IConfigurationValidator {
   async validate(config: IConfiguration): Promise<ValidatorResults> {
@@ -19,3 +21,5 @@ export class VimrcValidator implements IConfigurationValidator {
     // no-op
   }
 }
+
+configurationValidator.registerValidator(new VimrcValidator());
