+++ test/testUtils.ts
@@ -4,6 +4,7 @@ import * as os from 'os';
 import { join } from 'path';
 import { promisify } from 'util';
 import * as vscode from 'vscode';
+import * as path from 'path';
 
 import { Configuration } from './testConfiguration';
 import { Globals } from '../src/globals';
@@ -13,6 +14,33 @@ import { TextEditor } from '../src/textEditor';
 import { getAndUpdateModeHandler } from '../extension';
 import { commandLine } from '../src/cmd_line/commandLine';
 import { StatusBar } from '../src/statusBar';
+
+class TestMemento implements vscode.Memento {
+  private mapping = new Map<string, any>();
+  get<T>(key: string): T | undefined;
+  get<T>(key: string, defaultValue: T): T;
+  get(key: any, defaultValue?: any) {
+    return this.mapping.get(key) || defaultValue;
+  }
+
+  async update(key: string, value: any): Promise<void> {
+    this.mapping.set(key, value);
+  }
+}
+export class TestExtensionContext implements vscode.ExtensionContext {
+  subscriptions: { dispose(): any }[] = [];
+  workspaceState: vscode.Memento = new TestMemento();
+  globalState: vscode.Memento = new TestMemento();
+  extensionPath: string = 'inmem:///test';
+
+  asAbsolutePath(relativePath: string): string {
+    return path.resolve(this.extensionPath, relativePath);
+  }
+
+  storagePath: string | undefined;
+  globalStoragePath: string;
+  logPath: string;
+}
 
 export function rndName(): string {
   return Math.random()
@@ -32,22 +60,22 @@ export async function createRandomDir() {
   return createDir(dirPath);
 }
 
-export async function createEmptyFile(path: string) {
-  await promisify(fs.writeFile)(path, '');
-  return path;
+export async function createEmptyFile(fsPath: string) {
+  await promisify(fs.writeFile)(fsPath, '');
+  return fsPath;
 }
 
-export async function createDir(path: string) {
-  await promisify(fs.mkdir)(path);
-  return path;
+export async function createDir(fsPath: string) {
+  await promisify(fs.mkdir)(fsPath);
+  return fsPath;
 }
 
-export async function removeFile(path: string) {
-  return promisify(fs.unlink)(path);
+export async function removeFile(fsPath: string) {
+  return promisify(fs.unlink)(fsPath);
 }
 
-export async function removeDir(path: string) {
-  return promisify(fs.rmdir)(path);
+export async function removeDir(fsPath: string) {
+  return promisify(fs.rmdir)(fsPath);
 }
 
 /**
